Password protection for video integrated.

+++ portlets/lecture2go-portlet/docroot/WEB-INF/src/de/uhh/l2g/plugins/service/impl/InstitutionLocalServiceImpl.java
@@ -114,7 +114,7 @@ public class InstitutionLocalServiceImpl extends InstitutionLocalServiceBaseImpl
 
 	public Map<String, String> getAllSortedAsTree(int begin, int end) throws SystemException {
 		Map<String, String> allInstitutions = new LinkedHashMap<String, String>();
-		List<Institution> einListAll = InstitutionFinderUtil.findAllSortedAsTree(1, 20);
+		List<Institution> einListAll = InstitutionFinderUtil.findAllSortedAsTree(begin, end);
 
 		for (Institution institution : einListAll) {
 			String id = "" + institution.getInstitutionId();
@@ -127,6 +127,10 @@ public class InstitutionLocalServiceImpl extends InstitutionLocalServiceBaseImpl
 
 	public int getMaxSortByParentId(long parentId) throws SystemException {
 		return InstitutionFinderUtil.findMaxSortByParent(parentId);
+	}
+
+	public int getLockingElements(long institutionId) throws SystemException {
+		return InstitutionFinderUtil.findLockingElements(institutionId);
 	}
 
 	public List<Institution> getInstitutionsFromLectureseriesIdsAndVideoIds(ArrayList<Long> lectureseriesIds, ArrayList<Long> videoIds) {
@@ -271,15 +275,23 @@ public class InstitutionLocalServiceImpl extends InstitutionLocalServiceBaseImpl
 	   public Institution deleteInstitution(long institutionId, ServiceContext serviceContext)
 		        throws PortalException, SystemException {
 
-		        //TODO: check if Institution is empty
-		        Institution institution = getInstitution(institutionId);
 
-		        resourceLocalService.deleteResource(serviceContext.getCompanyId(),
-		        		Institution.class.getName(), ResourceConstants.SCOPE_INDIVIDUAL,
-		        		institutionId);
+		   		Institution institution = getInstitution(institutionId);
 
-		        updateSort(institution, 0);
-		        institution = deleteInstitution(institutionId);
+		        if (getLockingElements(institutionId) < 1){
+			        //Check if Institution is empty, i.e. no  subfacilities, lecture series, videos, and members
+
+			        resourceLocalService.deleteResource(serviceContext.getCompanyId(),
+			        		Institution.class.getName(), ResourceConstants.SCOPE_INDIVIDUAL,
+			        		institutionId);
+
+			        updateSort(institution, 0);
+			        institution = deleteInstitution(institutionId);
+
+			        //Remove Entry from Link Table
+			        Institution_HostLocalServiceUtil.deleteEntriesByInstitution(institutionId, serviceContext);
+		        }
+		        else { System.out.println("Could not delte "+ institution.getName());}
 
 		        return institution;
